backend/requirements.txt
fastapi
uvicorn[standard]
httpx
anthropic
sqlmodel
python-jose[cryptography]
python-dotenv
pydantic
